requirements.txt
python-telegram-bot==20.7
yt-dlp>=2024.1.0
celery[redis]==5.3.6
redis==5.0.1
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.7
aiofiles==23.2.1
httpx==0.26.0
stripe==7.12.0
openai-whisper==20231117
ffmpeg-python==0.2.0
